Make generic version of Matrix4D (issue #125)
